+++ GenericNetworkLayerExample/GenericNetworkLayerExample/AppSetting.swift
@@ -1,0 +1,17 @@
+//
+//  AppSetting.swift
+//  GenericNetworkLayerExample
+//
+//  Created by Cyril Romanovsky on 7/18/20.
+//  Copyright © 2020 Cyril Romanovsky. All rights reserved.
+//
+
+import Foundation
+
+class AppSetting {
+    
+    static let shared = AppSetting()
+    
+    let baseURLStr = "https://api.exchangeratesapi.io/"
+
+}
