+++ GenericNetworkLayerExample/GenericNetworkLayerExample/Network/Enums/HttpHeaderValue.swift
@@ -1,0 +1,14 @@
+//
+//  HttpHeaderValue.swift
+//  GenericNetworkLayerExample
+//
+//  Created by Cyril Romanovsky on 7/18/20.
+//  Copyright © 2020 Cyril Romanovsky. All rights reserved.
+//
+
+import Foundation
+
+/// The value of the header-type (i.e. "application/json")
+enum HttpHeaderValue: String {
+    case json = "application/json"
+}
