+++ GenericNetworkLayerExample/GenericNetworkLayerExample/Network/NetworkService.swift
@@ -9,24 +9,6 @@
 import Foundation
 
 class NetworkService {
-    ///Used to set a`URLRequest`'s HTTP Method
-    enum HttpMethod: String {
-        case get = "GET"
-        case patch = "PATCH"
-        case post = "POST"
-        case put = "PUT"
-        case delete = "DELETE"
-    }
-    
-    /// Used when the endpoint requires a header-type (i.e. "content-type") be specified in the header
-    enum HttpHeaderType: String {
-        case contentType = "Content-Type"
-    }
-
-    /// The value of the header-type (i.e. "application/json")
-    enum HttpHeaderValue: String {
-        case json = "application/json"
-    }
 
     // Used to switch between live and Mock Data
     var dataLoader: NetworkLoader
@@ -34,13 +16,11 @@ class NetworkService {
     init(dataLoader: NetworkLoader = URLSession.shared) {
         self.dataLoader = dataLoader
     }
-
     
     /// Create a request given a URL and requestMethod (get, post, create, etc...)
     func createRequest(
         url: URL?, method: HttpMethod,
-        headerType: HttpHeaderType? = nil,
-        headerValue: HttpHeaderValue? = nil,
+        headerVals: [HttpHeaderType: HttpHeaderValue]? = nil,
         params: [String: Any]? = nil
     ) -> URLRequest? {
         guard let requestUrl = url else {
@@ -53,9 +33,10 @@ class NetworkService {
             let jsonData = try? JSONSerialization.data(withJSONObject: requestParams)
             request.httpBody = jsonData
         }
-        if let headerType = headerType,
-            let headerValue = headerValue {
-            request.setValue(headerValue.rawValue, forHTTPHeaderField: headerType.rawValue)
+        if let headerType = headerVals {
+            headerType.forEach { (val) in
+                request.setValue( val.value.rawValue, forHTTPHeaderField: val.key.rawValue)
+            }
         }
         return request
     }
