+++ GenericNetworkLayerExample/GenericNetworkLayerExample/ViewController.swift
@@ -13,11 +13,13 @@ class ViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         let nl = NetworkLayer()
-        guard let url = URL(string: "https://api.exchangeratesapi.io/latest") else { return } // next step add url builder
-        nl.request(url: url) { (result: Result<(model: SuccussModel?, error: ErrorModel?), Error>) in
+        let requestData = Requests.getAllCurrency.requestData
+        
+        guard let url = URL(string: requestData.url) else { return }
+        nl.request(url: url, type: requestData.type) { (result: Result<(model: SuccussModel?, error: ErrorModel?), Error>) in
             switch result {
-            case .success(let succ):
-                print(succ)
+            case .success(let success):
+                print(success)
             case .failure(let error):
                 print(error)
             }
